requirements.txt
Django>=5.2
mysqlclient
django-allauth
PyJWT
Pillow
gunicorn
whitenoise